requirements.txt
anthropic>=0.39.0
openai>=1.50.0
pypdf>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.40.0
